Document workaround for the system clock falling behind
After discussing with VirtualBox developers, it seems that the TSC stops
on some older processors when the CPU is idle. On real hardware, the
kernel can detect this and will mark the tsc clocksource as unreliable,
but it can't do that when running in a VM, since VirtualBox 'fixes' the
clock. The workaround is to set the paravirtualization interface to none
on such systems.